Move Country to domain models
