+++ src/backend/TrafficCourts/Common/Features/Lookups/CountryLookup.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using TrafficCourts.Common.Models;
+using TrafficCourts.Domain.Models;
 
 namespace TrafficCourts.Common.Features.Lookups;
