--- src/backend/TrafficCourts/Common/Models/Country.cs
@@ -1,9 +1,0 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-namespace TrafficCourts.Common.Models;
-
-[ExcludeFromCodeCoverage]
-public record Country(
-    string CtryId, 
-    string CtryLongNm
-);
