+++ src/backend/TrafficCourts/TrafficCourts.Core/Domain/Models/Country.cs
@@ -1,0 +1,9 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace TrafficCourts.Domain.Models;
+
+[ExcludeFromCodeCoverage]
+public record Country(
+    string CtryId, 
+    string CtryLongNm
+);
